requirements.txt
streamlit
langchain
langchain-groq
langchain-google-genai
faiss-cpu
pypdf
python-dotenv
langchain_community